safari-600.7-branch/LayoutTests/svg/custom/svg-parse-overflow-5.html

<html>
<body>
<script>
document.body.offsetTop;
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<svg>
<foreignObject y="105.0e5000000000000000000000000000000000000000000000000000" transform="scale(5) skewX(5)">
</svg>
</body>
