safari-600.7-branch/LayoutTests/svg/custom/svg-parse-overflow-5-expected.txt

CONSOLE MESSAGE: line 9: Error: Invalid value for <foreignObject> attribute y="105.0e5000000000000000000000000000000000000000000000000000"
